\name{annotateTRM}
\alias{annotateTRM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Annotate a network object with information about clusters.
}
\description{
This function takes a network object and includes cluster information as piecolor attribute, suitable to be plotted with plotTRM()
}
\usage{
annotateTRM(g, target)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{g}{a network object.}
  \item{target}{target node (from findTRM())}
}
\author{
Diego Diez
}
