\name{biogrid_hs}
\docType{data}
\alias{biogrid_hs}
\title{Network dataset of class 'igraph'}
\description{
  Human protein-protein interaction (PPI) dataset from the BioGRID database release \Sexpr{require(igraph);data(biogrid_hs);biogrid_hs$info$version}.
}
\usage{
data(biogrid_hs)
}
\format{
	An igraph object.
}
\author{Diego Diez}
\keyword{datasets}

