\name{biogrid_mm}
\docType{data}
\alias{biogrid_mm}
\title{Network dataset of class 'igraph'}
\description{
  Mouse protein-protein interaction (PPI) dataset from the BioGRID database \Sexpr{require(igraph);data(biogrid_mm);biogrid_mm$info$version}.
}
\usage{
data(biogrid_mm)
}
\format{
	An igraph object.
}
\author{Diego Diez}
\keyword{datasets}

