\name{getLargestComp}
\alias{getLargestComp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Gets the largest connected component
}
\description{
Returns the largest connected component from a graph.}
\usage{
getLargestComp(g)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{g}{an igraph object.}
}

\author{
Diego Diez
}