\name{getMotifsFromEntrezgene}
\alias{getMotifsFromEntrezgene}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Retrieve PWMs associated with genes provided as entrezgene identifiers.
}
\description{
Retrieve PWMs associated with genes provided as entrezgene identifiers.
}
\usage{
getMotifsFromEntrezgene(e, organism)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{e}{
vector of entrezgene identifiers to retrieve exiting PWMs.
}
  \item{organism}{
target organism.
}
}
\author{
Diego Diez
}
