\name{getMotifsFromSymbol}
\alias{getMotifsFromSymbol}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Retrieve PWMs associated with genes provided as symbol.
}
\description{
Retrieve PWMs associated with genes provided as symbol.
}
\usage{
getMotifsFromSymbol(s, organism)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{s}{
vector of gene symbols.
}
  \item{organism}{
target organism.
}
}
\author{
Diego Diez
}