\name{getOrthologsFromBiomart}
\alias{getOrthologsFromBiomart}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Returns ortholog genes for a target organism
}
\description{
Returns ortholog genes for a target organism
}
\usage{
getOrthologsFromBiomart(eg, target_org, mart)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{eg}{
list of entrezgene ids to obtain orthologs.
}
  \item{target_org}{
target organism.
}
  \item{mart}{
mart object.
}
}

\author{
Diego Diez
}
