\name{getTFclass}
\alias{getTFclass}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Return the ontology in the TFclass database associated with an entrezgene identifier
}
\description{
Return the ontology in the TFclass database associated with an entrezgene identifier.
}
\usage{
getTFclass(dbname = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dbname}{SQLite file to use as database.}
}
\author{
Diego Diez
}