\name{plotDegree}
\alias{plotDegree}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot degree distribution for network nodes
}
\description{
Plots the degree distribution and fits a power law, returning in the legend the values of the fitted parameters.}
\usage{
plotDegree(g)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{g}{
igraph object
}
}

\author{
Diego Diez
}