\name{rTRM-package}
\alias{rTRM-package}
\alias{rTRM}
\docType{package}
\title{
	Identification transcription regulatory modules (TRMs)
}
\description{
	This package integrates transcription factor binding, enrichment analysis,
	gene expression, functional information  and protein-protein interaction
	data, and identifies transcription factor modules (TFM) that regulate
	cellular responses.
}
\details{
\tabular{ll}{
Package: \tab rTRM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-01\cr
License: \tab GPL-3\cr
}
%%~~ An overview of how to use the package, including the most important ~~
%%~~ functions ~~
}
\author{
Diego Diez

Maintainer: Diego Diez <diego10ruiz@gmail.com>
}

\keyword{ package }
\examples{
getAnnotations()
}
